#28 check the solution execution

+++ unarchive_from_url_param/test.sh
@@ -12,12 +12,18 @@ fi
 # turn on the environment
 vagrant up
 
-# check and execute the solution
-ansible-playbook playbook-servers.yml -i hosts
+# execute the solution
+tmpfile=$(mktemp)
+ansible-playbook playbook-servers.yml -i hosts | tee -a ${tmpfile}
+success=$(tail -4 ${tmpfile} | grep -c "failed=0")
+if [ ${success} -ne 2 ]; then
+	echo "Error on playbook-servers.yml execution."
+	exit 1
+fi
 
 # validate the solution
 ansible servers -i hosts -m shell -a "ls /var/target"
 
 # turn off the environment
 vagrant destroy -f
-rm -rf .vagrant/
+rm -rf .vagrant/ "$tmpfile"
